pyproject.toml
[project]
name = "video-util"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "google-genai>=1.3.0",
    "matplotlib>=3.10.1",
    "natsort>=8.4.0",
    "ollama>=0.4.7",
    "openai>=1.65.3",
    "opencv-python>=4.11.0.86",
    "pytesseract>=0.3.13",
    "python-dotenv>=1.0.1",
    "streamlit>=1.42.2",
]
